pyproject.toml
[tool.poetry]
name = "repository"
version = "0.1.0"
description = ""
authors = ["max31ru12 <maxevg72@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic-marshals = "^0.3.12"
passlib = "^1.7.4"
httpx = "^0.27.0"
faker = "^24.8.0"
alembic = "^1.13.1"
trio = "^0.25.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.6"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
